Reorganize zone_connection/social.rs into several files
