+++ servers/world/src/zone_connection/mod.rs
@@ -42,8 +42,12 @@ mod actors;
 mod chat;
 mod effect;
 mod event;
+mod friends;
 mod item;
+mod linkshell;
 mod lua;
+mod mail;
+mod party;
 mod quest;
 mod social;
 pub mod spawn_allocator;
